refactor: vertically organize ChainSync
