+++ src/Hoard/ChainSync/Listeners.hs
@@ -1,35 +1,38 @@
-module Hoard.Listeners.ChainSyncEventListener
-    ( chainSyncHeaderReceivedListener
-    , chainSyncStartedListener
-    , chainSyncRollBackwardListener
-    , chainSyncRollForwardListener
-    , chainSyncIntersectionFoundListener
+module Hoard.ChainSync.Listeners
+    ( chainSyncHeaderReceived
+    , chainSyncStarted
+    , chainSyncRollBackward
+    , chainSyncRollForward
+    , chainSyncIntersectionFound
     ) where
 
+import Cardano.Api.LedgerState ()
 import Effectful (Eff, (:>))
+import Ouroboros.Consensus.Block (BlockNo (..), SlotNo (..))
+import Ouroboros.Consensus.Block.Abstract (blockNo, blockSlot)
 
-import Hoard.Data.Header (Header (..))
-import Hoard.Data.Header.Extract (extractHeaderData)
-import Hoard.Effects.HeaderRepo (HeaderRepo, upsertHeader)
-import Hoard.Effects.Log (Log)
-import Hoard.Effects.Log qualified as Log
-import Hoard.Network.Events
+import Hoard.ChainSync.Events
     ( ChainSyncIntersectionFound (..)
     , ChainSyncStarted (..)
     , HeaderReceived (..)
     , RollBackward (..)
     , RollForward (..)
     )
+import Hoard.Data.BlockHash (blockHashFromHeader)
+import Hoard.Data.Header (Header (..))
+import Hoard.Effects.HeaderRepo (HeaderRepo, upsertHeader)
+import Hoard.Effects.Log (Log)
+import Hoard.Effects.Log qualified as Log
 
 
 -- | Listener that handles ChainSync HeaderReceived events
 --
 -- Extracts header data and persists it to the database.
-chainSyncHeaderReceivedListener
+chainSyncHeaderReceived
     :: (Log :> es, HeaderRepo :> es)
     => HeaderReceived
     -> Eff es ()
-chainSyncHeaderReceivedListener event = do
+chainSyncHeaderReceived event = do
     Log.debug "📦 Header received!"
     -- Extract header data from the event
     let header = extractHeaderData event
@@ -39,36 +42,47 @@ chainSyncHeaderReceivedListener event = do
 
 
 -- | Listener that handles ChainSync started events
-chainSyncStartedListener
+chainSyncStarted
     :: (Log :> es)
     => ChainSyncStarted
     -> Eff es ()
-chainSyncStartedListener event = do
+chainSyncStarted event = do
     Log.info $ "⛓️  ChainSync protocol started at " <> show event.timestamp
 
 
 -- | Listener that handles ChainSync rollback events
-chainSyncRollBackwardListener
+chainSyncRollBackward
     :: (Log :> es)
     => RollBackward
     -> Eff es ()
-chainSyncRollBackwardListener _event = do
+chainSyncRollBackward _event = do
     Log.info "⏪ Rollback occurred"
 
 
 -- | Listener that handles ChainSync roll forward events
-chainSyncRollForwardListener
+chainSyncRollForward
     :: (Log :> es)
     => RollForward
     -> Eff es ()
-chainSyncRollForwardListener _event = do
+chainSyncRollForward _event = do
     Log.info "⏩ RollForward occurred"
 
 
 -- | Listener that handles ChainSync intersection found events
-chainSyncIntersectionFoundListener
+chainSyncIntersectionFound
     :: (Log :> es)
     => ChainSyncIntersectionFound
     -> Eff es ()
-chainSyncIntersectionFoundListener _event = do
+chainSyncIntersectionFound _event = do
     Log.info "🎯 ChainSync intersection found"
+
+
+-- | Extract header data from a HeaderReceivedData event
+extractHeaderData :: HeaderReceived -> Header
+extractHeaderData event =
+    Header
+        { hash = blockHashFromHeader event.header
+        , slotNumber = unSlotNo $ blockSlot event.header
+        , blockNumber = unBlockNo $ blockNo event.header
+        , firstSeenAt = event.timestamp
+        }
