+++ src/Hoard/Collector.hs
@@ -16,6 +16,7 @@ import Effectful.Timeout (Timeout)
 import Prelude hiding (Reader, State, asks, gets, modify, state)
 
 import Hoard.Bootstrap (bootstrapPeers)
+import Hoard.ChainSync.Events (HeaderReceived (..))
 import Hoard.Control.Exception (isGracefulShutdown, withExceptionLogging)
 import Hoard.Data.BlockHash (blockHashFromHeader)
 import Hoard.Data.Peer (Peer (..))
@@ -36,11 +37,7 @@ import Hoard.Effects.Output (Output, output, runOutputChan)
 import Hoard.Effects.PeerRepo (PeerRepo, getAllPeers, updatePeerFailure)
 import Hoard.Effects.Pub (Pub, publish)
 import Hoard.Events.Collector (CollectorEvent (..))
-import Hoard.Network.Events
-    ( BlockFetchRequest (..)
-    , BlockReceived (..)
-    , HeaderReceived (..)
-    )
+import Hoard.Network.Events (BlockFetchRequest (..), BlockReceived (..))
 import Hoard.Types.Environment (Config (..))
 import Hoard.Types.HoardState (BlocksBeingFetched (..), HoardState (..), connectedPeers)
