+++ src/Hoard/Effects/NodeToNode/Config.hs
@@ -1,0 +1,13 @@
+module Hoard.Effects.NodeToNode.Config (Config (..)) where
+
+import Hoard.ChainSync.Events (HeaderReceived (..))
+import Hoard.Data.Peer (Peer (..))
+import Hoard.Network.Events (BlockFetchRequest (..), BlockReceived (..))
+
+
+data Config m = Config
+    { awaitBlockFetchRequests :: m (NonEmpty BlockFetchRequest)
+    , emitFetchedHeader :: HeaderReceived -> m ()
+    , emitFetchedBlock :: BlockReceived -> m ()
+    , peer :: Peer
+    }
