+++ src/Hoard/Types/Cardano.hs
@@ -4,10 +4,17 @@ module Hoard.Types.Cardano
     , CardanoHeader
     , CardanoPoint
     , CardanoTip
+    , CardanoCodecs
     ) where
 
+import Cardano.Api ()
+import Codec.CBOR.Read (DeserialiseFailure)
+import Data.ByteString.Lazy qualified as LBS
+import Network.Socket (SockAddr)
 import Ouroboros.Consensus.Cardano.Block qualified as Consensus
+import Ouroboros.Consensus.Network.NodeToNode (Codecs (..))
 import Ouroboros.Network.Block qualified as Network
+import Prelude hiding (Reader, State, asks, evalState, gets)
 
 
 -- We use StandardCrypto which is the standard cryptographic primitives used in
@@ -29,3 +36,18 @@ type CardanoPoint = Network.Point CardanoBlock
 
 
 type CardanoTip = Network.Tip CardanoBlock
+
+
+type CardanoCodecs =
+    Codecs
+        CardanoBlock
+        SockAddr
+        DeserialiseFailure
+        IO
+        LBS.ByteString
+        LBS.ByteString
+        LBS.ByteString
+        LBS.ByteString
+        LBS.ByteString
+        LBS.ByteString
+        LBS.ByteString
